+++ src/PresentationalComponents/TableView/TableFooter.js
@@ -1,27 +1,33 @@
-import { Pagination, PaginationVariant } from '@patternfly/react-core';
-import { TableToolbar } from '@redhat-cloud-services/frontend-components/TableToolbar';
+import { Pagination, PaginationVariant, Skeleton } from '@patternfly/react-core';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-const TableFooter = ({ page, perPage, onSetPage, totalItems, onPerPageSelect, paginationOUIA }) => {
+const TableFooter = ({ isLoading, page, perPage, onSetPage, totalItems, onPerPageSelect, paginationOUIA }) => {
     return (
-        <TableToolbar>
-            <Pagination
-                itemCount={totalItems}
-                perPage={perPage}
-                page={page}
-                onSetPage={onSetPage}
-                onPerPageSelect={onPerPageSelect}
-                widgetId={`pagination-options-menu-bottom`}
-                variant={PaginationVariant.bottom}
-                ouiaId={paginationOUIA}
-                isDisabled={totalItems === 0}
-            />
-        </TableToolbar>
+        <>
+            {isLoading ? (
+                <div className="pf-c-pagination pf-m-bottom">
+                    <Skeleton fontSize="xl" width="350px" style={{ margin: 10 }} />
+                </div>
+            ) :
+                <Pagination
+                    itemCount={totalItems}
+                    perPage={perPage}
+                    page={page}
+                    onSetPage={onSetPage}
+                    onPerPageSelect={onPerPageSelect}
+                    widgetId={`pagination-options-menu-bottom`}
+                    variant={PaginationVariant.bottom}
+                    ouiaId={paginationOUIA}
+                    isDisabled={totalItems === 0}
+                />
+            }
+        </>
     );
 };
 
 TableFooter.propTypes = {
+    isLoading: PropTypes.bool,
     onSetPage: PropTypes.func,
     onPerPageSelect: PropTypes.func,
     page: PropTypes.number,