+++ src/store/index.js
@@ -16,7 +16,9 @@ import { GlobalFilterStore } from './Reducers/GlobalFilterStore';
 import { PatchSetsReducer } from './Reducers/PatchSetsReducer';
 import { PatchSetDetailStore } from './Reducers/PatchSetDetailStore';
 import { SpecificPatchSetReducer } from './Reducers/SpecificPatchSetReducer';
-import { legacy_createStore as createStore, applyMiddleware, combineReducers } from 'redux';
+import { legacy_createStore as createStore, applyMiddleware, combineReducers, compose } from 'redux';
+
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
 export const defaultReducers = {
     AdvisoryListStore,
@@ -37,4 +39,7 @@ export const defaultReducers = {
     notifications: notificationsReducer
 };
 
-export const store = createStore(combineReducers(defaultReducers), applyMiddleware(promiseMiddleware, notificationsMiddleware()));
+export const store = createStore(
+    combineReducers(defaultReducers),
+    composeEnhancers(applyMiddleware(promiseMiddleware, notificationsMiddleware()))
+);